RealTimeWeatherStation using RPI/requirements.txt
python-dotenv
firebase-admin
Adafruit_DHT
Adafruit_BMP
requests